Some cleanup and godoc
